storage: move health_operator to the healthcheck module
