+++ src/storage/src/lib.rs
@@ -88,7 +88,7 @@ pub mod source;
 pub mod statistics;
 pub mod storage_state;
 
-mod healthcheck;
+pub(crate) mod healthcheck;
 
 pub use decode::metrics::DecodeMetrics;
 pub use server::{serve, Config, Server};
